frontend/requirements.txt
streamlit
requests
tensorflow
numpy
pillow
openai
python-dotenv
